Moved checkers to their own folder.
