+++ tools/e2etesting/TestEventProcessor/TestEventProcessor.Businesslogic/Checkers/MissingTimestampsChecker.cs
@@ -3,7 +3,7 @@
 //  Licensed under the MIT License (MIT). See License.txt in the repo root for license information.
 // ------------------------------------------------------------
 
-namespace TestEventProcessor.BusinessLogic {
+namespace TestEventProcessor.BusinessLogic.Checkers {
     using Microsoft.Extensions.Logging;
     using System;
     using System.Collections.Generic;
